refactor(api-reference): rename file to kebab case
